# docs/swagger.yaml
definitions:
  internal_controllers.AssignMenuRequest:
    properties:
      can_create:
        type: boolean
      can_delete:
        type: boolean
      can_edit:
        type: boolean
      can_view:
        type: boolean
    type: object
...
  internal_controllers.CreateMenuRequest:
    properties:
      icon:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      parent_id:
        type: string
      path:
        type: string
      sort_order:
        type: integer
    required:
    - name
    type: object
...
  internal_controllers.UpdateMenuRequest:
    properties:
      icon:
        type: string
      is_active:
        type: boolean
      name:
        maxLength: 100
        minLength: 2
        type: string
      parent_id:
        type: string
      path:
        type: string
      sort_order:
        type: integer
    required:
    - name
    type: object
...
  user-management_internal_models.LoginRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  user-management_internal_models.LoginResponse:
    properties:
      expires_at:
        type: string
      token:
        type: string
      user:
        $ref: '#/definitions/user-management_internal_models.UserResponse'
    type: object
...
  user-management_internal_models.UserResponse:
    properties:
      created_at:
        type: string
      department:
        type: string
      email:
        type: string
      is_active:
        type: boolean
      last_login_at:
        type: string
      name:
        type: string
      nik:
        type: string
      phone:
        type: string
      position:
        type: string
      role:
        type: string
      user_id:
        type: string
      username:
        type: string
    type: object
  user-management_internal_utils.Response:
    properties:
      code:
        type: string
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
info:
  contact: {}
paths:
...
  /api/v1/auth/history:
    get:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - default: 20
        description: Limit
        in: query
        name: limit
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user-management_internal_utils.Response'
      security:
      - ApiKeyAuth: []
      summary: Get my login history
      tags:
      - Login History
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user-management_internal_models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/user-management_internal_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/user-management_internal_models.LoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/user-management_internal_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/user-management_internal_utils.Response'
      summary: User login
      tags:
      - Auth
  /api/v1/auth/logout:
    post:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user-management_internal_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/user-management_internal_utils.Response'
      security:
      - ApiKeyAuth: []
      summary: User logout
      tags:
      - Auth
  /api/v1/auth/me:
    get:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/user-management_internal_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/user-management_internal_models.UserResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/user-management_internal_utils.Response'
      security:
      - ApiKeyAuth: []
      summary: Get current user info
      tags:
      - Auth
...
  /api/v1/login-history:
    get:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: page_size
        type: integer
      - description: Search by email
        in: query
        name: search
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user-management_internal_utils.Response'
      security:
      - ApiKeyAuth: []
      summary: Get all login history
      tags:
      - Login History
  /api/v1/menus:
    get:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user-management_internal_utils.Response'
      security:
      - ApiKeyAuth: []
      summary: Get all menus
      tags:
      - Menus
    post:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Menu data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_controllers.CreateMenuRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/user-management_internal_utils.Response'
      security:
      - ApiKeyAuth: []
      summary: Create menu
      tags:
      - Menus
  /api/v1/menus/{id}:
    delete:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Menu ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user-management_internal_utils.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete menu
      tags:
      - Menus
    put:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Menu ID
        in: path
        name: id
        required: true
        type: string
      - description: Menu data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_controllers.UpdateMenuRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user-management_internal_utils.Response'
      security:
      - ApiKeyAuth: []
      summary: Update menu
      tags:
      - Menus
  /api/v1/menus/tree:
    get:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user-management_internal_utils.Response'
      security:
      - ApiKeyAuth: []
      summary: Get menu tree
      tags:
      - Menus
  /api/v1/menus/user:
    get:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user-management_internal_utils.Response'
      security:
      - ApiKeyAuth: []
      summary: Get user menus
      tags:
      - Menus
...
  /api/v1/roles/{roleId}/menus:
    get:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Role ID
        in: path
        name: roleId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user-management_internal_utils.Response'
      security:
      - ApiKeyAuth: []
      summary: Get menus by role
      tags:
      - Menus
  /api/v1/roles/{roleId}/menus/{menuId}:
    delete:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Role ID
        in: path
        name: roleId
        required: true
        type: string
      - description: Menu ID
        in: path
        name: menuId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user-management_internal_utils.Response'
      security:
      - ApiKeyAuth: []
      summary: Remove menu from role
      tags:
      - Menus
    post:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Role ID
        in: path
        name: roleId
        required: true
        type: string
      - description: Menu ID
        in: path
        name: menuId
        required: true
        type: string
      - description: Menu permissions
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_controllers.AssignMenuRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user-management_internal_utils.Response'
      security:
      - ApiKeyAuth: []
      summary: Assign menu to role
      tags:
      - Menus
...
swagger: "2.0"